requirements.txt
python-dotenv~=1.1.0
pydantic~=2.12.5
langgraph~=1.0.5